requirements.txt
aiogram
httpx
motor
python-dotenv